services/ai-service/requirements.txt
# ── OpenSalesAI AI Service — Python Dependencies ──

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
orjson==3.10.13
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis[hiredis]==5.2.1

# Vector Database
qdrant-client==1.12.1

# LLM / AI
langchain==0.3.14
langchain-core==0.3.30
langchain-community==0.3.14
langgraph==0.2.62
httpx==0.28.1

# Embeddings
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.47.1

# ML Models
scikit-learn==1.6.0
xgboost==2.1.3
prophet==1.1.6
ortools==9.11.4210
pandas==2.2.3
numpy==2.2.1

# Speech
faster-whisper==1.1.0
gTTS==2.5.4
pydub==0.25.1

# Auth
PyJWT==2.10.1
cryptography==44.0.0

# Template Engine
Jinja2==3.1.5

# Monitoring
prometheus-client==0.21.1

# Object Storage
minio==7.2.12

# MLflow
mlflow==2.19.0

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0

# Development / Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
ruff==0.8.6
black==24.10.0
mypy==1.14.1
